This revision includes minor changes: C-drizzle version updated, WCSMap object now has all shift info as attributes, and a new interface has been added to run MultiDrizzle from the command-line using the mdriz module. WJH
git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/development/trunk/BigBlackBox@7876 fe389314-cf27-0410-b35b-8c050e845b92

+++ lib/imageObject.py
@@ -482,7 +482,7 @@ class baseImageObject:
     def _buildMask(self,dqarr,bitvalue):
         """ Builds a bit-mask from an input DQ array and a bitvalue flag"""
         if bitvalue == None:
-            return (dqarr * 0.0) + 1.0
+            return ((dqarr * 0.0) + 1.0).astype(np.uint8)
         _maskarr = np.bitwise_or(dqarr,np.array([bitvalue]))
         return np.choose(np.greater(_maskarr,bitvalue),(1,0)).astype(np.uint8)
